+++ deltachat/message_test.go
@@ -8,7 +8,7 @@ import (
 
 func TestMessage(t *testing.T) {
 	t.Parallel()
-	acc := acfactory.GetOnlineAccount()
+	acc := acfactory.OnlineAccount()
 	defer acc.Manager.Rpc.Stop()
 
 	chat, err := acc.Me().CreateChat()
@@ -40,7 +40,7 @@ func TestMessage(t *testing.T) {
 
 func TestMessage_WebxdcInfo(t *testing.T) {
 	t.Parallel()
-	acc := acfactory.GetOnlineAccount()
+	acc := acfactory.OnlineAccount()
 	defer acc.Manager.Rpc.Stop()
 
 	chat, err := acc.Me().CreateChat()
@@ -51,7 +51,7 @@ func TestMessage_WebxdcInfo(t *testing.T) {
 	info, err := msg.WebxdcInfo()
 	assert.NotNil(t, err)
 
-	msg, err = chat.SendMsg(MsgData{Text: "testing webxdc", File: acfactory.GetTestWebxdc()})
+	msg, err = chat.SendMsg(MsgData{Text: "testing webxdc", File: acfactory.TestWebxdc()})
 	assert.Nil(t, err)
 
 	info, err = msg.WebxdcInfo()
@@ -61,7 +61,7 @@ func TestMessage_WebxdcInfo(t *testing.T) {
 
 func TestMessage_SendMsg(t *testing.T) {
 	t.Parallel()
-	acc := acfactory.GetOnlineAccount()
+	acc := acfactory.OnlineAccount()
 	defer acc.Manager.Rpc.Stop()
 
 	chat, err := acc.Me().CreateChat()
@@ -70,36 +70,36 @@ func TestMessage_SendMsg(t *testing.T) {
 	_, err = chat.SendMsg(MsgData{Location: &[2]float64{1, 1}})
 	assert.Nil(t, err)
 
-	WaitForEvent(acc, EventLocationChanged{})
+	acfactory.WaitForEvent(acc, EventLocationChanged{})
 
-	acc2 := acfactory.GetOnlineAccount()
+	acc2 := acfactory.OnlineAccount()
 	defer acc2.Manager.Rpc.Stop()
 	chat, err = acfactory.CreateChat(acc2, acc)
 	acc.SetConfig("delete_server_after", "1")
 	_, err = chat.SendMsg(MsgData{Text: "test"})
 	assert.Nil(t, err)
-	WaitForEvent(acc, EventImapMessageDeleted{})
+	acfactory.WaitForEvent(acc, EventImapMessageDeleted{})
 }
 
 func TestMessage_StatusUpdates(t *testing.T) {
 	t.Parallel()
-	acc1 := acfactory.GetOnlineAccount()
+	acc1 := acfactory.OnlineAccount()
 	defer acc1.Manager.Rpc.Stop()
-	acc2 := acfactory.GetOnlineAccount()
+	acc2 := acfactory.OnlineAccount()
 	defer acc2.Manager.Rpc.Stop()
 
 	chat1, err := acfactory.CreateChat(acc1, acc2)
 	assert.Nil(t, err)
 
-	msg, err := chat1.SendMsg(MsgData{Text: "testing webxdc", File: acfactory.GetTestWebxdc()})
+	msg, err := chat1.SendMsg(MsgData{Text: "testing webxdc", File: acfactory.TestWebxdc()})
 	assert.Nil(t, err)
-	snapshot, err := acfactory.GetNextMsg(acc2)
+	snapshot, err := acfactory.NextMsg(acc2)
 	assert.Nil(t, err)
 
 	assert.Nil(t, msg.SendStatusUpdate(`{"payload": "test payload"}`, "update 1"))
-	WaitForEvent(acc2, EventWebxdcStatusUpdate{})
+	acfactory.WaitForEvent(acc2, EventWebxdcStatusUpdate{})
 	assert.Nil(t, msg.Delete())
-	WaitForEvent(acc1, EventWebxdcInstanceDeleted{})
+	acfactory.WaitForEvent(acc1, EventWebxdcInstanceDeleted{})
 
 	msg = &Message{acc2, snapshot.Id}
 	updates, err := msg.StatusUpdates(0)
@@ -109,9 +109,9 @@ func TestMessage_StatusUpdates(t *testing.T) {
 
 func TestMessage_ContinueAutocryptKeyTransfer(t *testing.T) {
 	t.Parallel()
-	acc1 := acfactory.GetOnlineAccount()
+	acc1 := acfactory.OnlineAccount()
 	defer acc1.Manager.Rpc.Stop()
-	acc2 := acfactory.GetUnconfiguredAccount()
+	acc2 := acfactory.UnconfiguredAccount()
 	defer acc2.Manager.Rpc.Stop()
 
 	addr, err := acc1.GetConfig("configured_addr")
@@ -130,7 +130,7 @@ func TestMessage_ContinueAutocryptKeyTransfer(t *testing.T) {
 
 	selfchat, err := acc2.Me().CreateChat()
 	assert.Nil(t, err)
-	event := waitForEvent(acc2, EventMsgsChanged{}, selfchat.Id).(EventMsgsChanged)
+	event := acfactory.WaitForEventInChat(acc2, EventMsgsChanged{}, selfchat.Id).(EventMsgsChanged)
 	assert.NotEmpty(t, event.MsgId)
 	msg := &Message{acc2, event.MsgId}
 	assert.Nil(t, msg.ContinueAutocryptKeyTransfer(code))
@@ -138,11 +138,11 @@ func TestMessage_ContinueAutocryptKeyTransfer(t *testing.T) {
 
 func TestMsgSnapshot_ParseMemberAddedRemoved(t *testing.T) {
 	t.Parallel()
-	acc1 := acfactory.GetOnlineAccount()
+	acc1 := acfactory.OnlineAccount()
 	defer acc1.Manager.Rpc.Stop()
 	addr1, err := acc1.GetConfig("configured_addr")
 	assert.Nil(t, err)
-	acc2 := acfactory.GetOnlineAccount()
+	acc2 := acfactory.OnlineAccount()
 	defer acc2.Manager.Rpc.Stop()
 	addr2, err := acc2.GetConfig("configured_addr")
 	assert.Nil(t, err)
@@ -164,7 +164,7 @@ func TestMsgSnapshot_ParseMemberAddedRemoved(t *testing.T) {
 	// promote group
 	msg, _ := chat1.SendText("test")
 	for {
-		event := waitForEvent(acc1, EventMsgsChanged{}, chat1.Id).(EventMsgsChanged)
+		event := acfactory.WaitForEventInChat(acc1, EventMsgsChanged{}, chat1.Id).(EventMsgsChanged)
 		if event.MsgId == msg.Id {
 			break
 		}
@@ -177,7 +177,7 @@ func TestMsgSnapshot_ParseMemberAddedRemoved(t *testing.T) {
 	// add new member
 	assert.Nil(t, chat1.AddContact(contact2))
 	// acc1 side
-	event := waitForEvent(acc1, EventMsgsChanged{}, chat1.Id).(EventMsgsChanged)
+	event := acfactory.WaitForEventInChat(acc1, EventMsgsChanged{}, chat1.Id).(EventMsgsChanged)
 	assert.NotEmpty(t, event.MsgId)
 	msg = &Message{acc1, event.MsgId}
 	snapshot, err = msg.Snapshot()
@@ -190,7 +190,7 @@ func TestMsgSnapshot_ParseMemberAddedRemoved(t *testing.T) {
 	assert.Equal(t, acc1.Me(), actor)
 	assert.Equal(t, contact2, target)
 	// acc2 side
-	snapshot, err = acfactory.GetNextMsg(acc2)
+	snapshot, err = acfactory.NextMsg(acc2)
 	assert.Nil(t, err)
 	assert.Equal(t, SysmsgMemberAddedToGroup, snapshot.SystemMessageType)
 	actor, target, err = snapshot.ParseMemberAdded()
@@ -201,7 +201,7 @@ func TestMsgSnapshot_ParseMemberAddedRemoved(t *testing.T) {
 	// remove new member
 	assert.Nil(t, chat1.RemoveContact(contact3acc1))
 	// acc1 side
-	event = waitForEvent(acc1, EventMsgsChanged{}, chat1.Id).(EventMsgsChanged)
+	event = acfactory.WaitForEventInChat(acc1, EventMsgsChanged{}, chat1.Id).(EventMsgsChanged)
 	assert.NotEmpty(t, event.MsgId)
 	msg = &Message{acc1, event.MsgId}
 	snapshot, err = msg.Snapshot()
@@ -214,7 +214,7 @@ func TestMsgSnapshot_ParseMemberAddedRemoved(t *testing.T) {
 	assert.Equal(t, acc1.Me(), actor)
 	assert.Equal(t, contact3acc1, target)
 	// acc2 side
-	snapshot, err = acfactory.GetNextMsg(acc2)
+	snapshot, err = acfactory.NextMsg(acc2)
 	assert.Nil(t, err)
 	assert.Equal(t, SysmsgMemberRemovedFromGroup, snapshot.SystemMessageType)
 	actor, target, err = snapshot.ParseMemberRemoved()
@@ -225,7 +225,7 @@ func TestMsgSnapshot_ParseMemberAddedRemoved(t *testing.T) {
 	// leave
 	assert.Nil(t, chat1.Leave())
 	// acc1 side
-	event = waitForEvent(acc1, EventMsgsChanged{}, chat1.Id).(EventMsgsChanged)
+	event = acfactory.WaitForEventInChat(acc1, EventMsgsChanged{}, chat1.Id).(EventMsgsChanged)
 	assert.NotEmpty(t, event.MsgId)
 	msg = &Message{acc1, event.MsgId}
 	snapshot, err = msg.Snapshot()
@@ -236,7 +236,7 @@ func TestMsgSnapshot_ParseMemberAddedRemoved(t *testing.T) {
 	assert.Equal(t, acc1.Me(), actor)
 	assert.Equal(t, acc1.Me(), target)
 	// acc2 side
-	snapshot, err = acfactory.GetNextMsg(acc2)
+	snapshot, err = acfactory.NextMsg(acc2)
 	assert.Nil(t, err)
 	assert.Equal(t, SysmsgMemberRemovedFromGroup, snapshot.SystemMessageType)
 	actor, target, err = snapshot.ParseMemberRemoved()